refactor(kernelsu): update ioctl request parameter to uint32_t and cast for type safety
This change modifies the `ksuctl` function to accept a `uint32_t` type for the request parameter, aligning with more precise type handling. The cast to `int` when invoking `ioctl` ensures compatibility across different system architectures. This refactor enhances code clarity and reduces potential type-related issues in kernel communication.

+++ zygiskd/src/root_impl/kernelsu.c
@@ -119,10 +119,10 @@ static int init_driver_fd(void) {
 }
 
 /* Execute ioctl command */
-static int ksuctl(unsigned long request, void *arg) {
+static int ksuctl(uint32_t request, void *arg) {
   if (driver_fd < 0) return -1;
 
-  int ret = ioctl(driver_fd, request, arg);
+  int ret = ioctl(driver_fd, (int)request, arg);
   if (ret < 0) {
     return -1;
   }
